requirements.txt
flask
openai
PyMuPDF
dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client